SFB-65: move editr format rules to code-editor folder
