+++ app/utils/code-editor/format/rules/add-new-line-after-dot.js
