+++ app/utils/code-editor/format/rules/add-tab-before-line-ending-with-dot-and-previous-with-semicolon.js
