+++ app/utils/code-editor/format/rules/add-tab-before-line-with-semicolon.js
