+++ app/utils/code-editor/format/rules/parse-line-when-ending-on-dot-and-semicolon-included.js
