+++ app/utils/code-editor/format/rules/remove-all-spaces-from-each-line.js
